+++ inst/tinytest/test-format_tt.R
@@ -4,8 +4,8 @@ using("tinysnapshot")
 
 # numeric vector input
 a <- c(98938272783457, 7288839482, 29111727, 93945)
-b <- format_tt(a, num_suffix = TRUE, digits = 1)
-expect_equivalent(b, c("98.9T", "7.3B", "29.1M", "93.9K"))
+b <- format_tt(a, num_suffix = TRUE, digits = 2)
+expect_equivalent(b, c("99T", "7.3B", "29M", "94K"))
 
 
 # num_suffix (vignette)