+++ NEWS.md
@@ -12,10 +12,12 @@
 Misc:
 
 * Improved support for `tibble`. ANSI characters (ex: fancy `pillar` formatting) are stripped automatically or converted to HTML when the `fansi` package is installed. `fansi` is a dependency of `tibble`, so it should often be installed.
+* New `tinytable_tt_digits` global option can set the default number of digits in the `tt()` function.
 
 Breaking changes:
 
 * In some cases, `format_tt()` could previously be applied sequentially to apply two formats to the same cell. Now, multiple calls to `format_tt()` can still be make chained with pipes, but they must apply to different cells with `i`, `j`, otherwise only the last change is respected. One exception is the `escape` argument which can be applied to pre-formatted cells.
+* The "tinytable_tabularray_placement" global option is renamed to "tabularray_tt_placement".
 
 ## 0.0.5